requirements.txt
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==4.6.3
Werkzeug==3.0.1
python-dotenv